chore(cne-99): separated class into files
For a more readable project structure, decided to separate classes into different files you can just call upon.
